Separate earley from parse() | add valuator to Rule
The earley function takes several parameters and spits out a set of
states (one for each token). This should be its own function since it
must be tested apart from the DFS tree constructor.

The Rule constructor takes an optional third parameter, the valuator,
which is the function used to evaluate the expression given values for
each symbol.