Fix the issue of multiple rules existing in same earley state
Duplicate earley items should not exist concurrently in a given earley
state, but duplicate _rules_ should be allowed. This means that if two
earley items have the same rule, but different origins or positions,
they should be allowed to exist in the same state.

This will allow certain ambiguous grammars/sentences to successfully
parse that would otherwise fail.