Create CFG class to contain a set of rules + spec
The CFG class is just like an Array, so it can be used with `parse()`
and other functions that take a list of rules as a grammar, except its
primary method, `rule()`, accepts either a `Rule` instance OR a string
of the form
    A -> B C D
which describes a production.

Symbols in stringy productions are matched with existing Sym objects in
the grammar (so now the name *does* matter).

Also, terminal symbols are distinguished by surrounding quotes "", '',
or bare regex constructors /.*?/g.